fix bf16

+++ bmtrain/csrc/cuda/cross_entropy.cu
@@ -1,4 +1,5 @@
-#include <cuda_fp16.h>
+// #include <cuda_fp16.h>
+#include <cuda_bf16.h>
 #include <torch/extension.h>
 #include <ATen/cuda/CUDAContext.h>
 #include "reduce.cuh"
@@ -7,9 +8,9 @@ namespace {
 // blocks <m>,      threads<1024>
 __global__ void cross_entropy_forward(
     int64_t n,
-    const half *input,      // (m, n)
+    const nv_bfloat16 *input,      // (m, n)
     const int32_t *target,  // (m)
-    half *softmax,          // (m, n)
+    nv_bfloat16 *softmax,          // (m, n)
     float *output,          // (m)
     int32_t ignore_index
 ) {
@@ -17,24 +18,24 @@ __global__ void cross_entropy_forward(
 
     float local_max = -INFINITY;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_max = fmaxf(__half2float(input[base_idx + i]), local_max);
+        local_max = fmaxf(__bfloat162float(input[base_idx + i]), local_max);
     }
 
     local_max = fmaxf(block_allreduce_max(local_max), -1e6);
     
     float local_sum = 0;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_sum += expf(__half2float(input[base_idx + i]) - local_max);
+        local_sum += expf(__bfloat162float(input[base_idx + i]) - local_max);
     }
     local_sum = block_allreduce_sum(local_sum) + 1e-10; // avoid nan
     
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        softmax[base_idx + i] = __float2half( expf(__half2float(input[base_idx + i]) - local_max) / local_sum );
+        softmax[base_idx + i] = __float2bfloat16( expf(__bfloat162float(input[base_idx + i]) - local_max) / local_sum );
     }
 
     if (threadIdx.x == 0) {
         if (target[blockIdx.x] != ignore_index) {
-            output[blockIdx.x] = -__half2float(input[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
+            output[blockIdx.x] = -__bfloat162float(input[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
         } else {
             output[blockIdx.x] = 0;
         }
@@ -46,21 +47,21 @@ __global__ void cross_entropy_backward(
     int64_t n,
     const float *grad_output,   // (m)
     const int32_t *target,      // (m)
-    const half *softmax,        // (m, n)
-    half *grad_input,           // (m, n)
+    const nv_bfloat16 *softmax,        // (m, n)
+    nv_bfloat16 *grad_input,           // (m, n)
     int32_t ignore_index
 ) {
     int64_t base_idx = blockIdx.x * n;
 
     int32_t t = target[blockIdx.x];
     if (t == ignore_index) {
-        half v = __float2half(0.);
+        nv_bfloat16 v = __float2bfloat16(0.);
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             grad_input[base_idx + i] = v;
         }
     }
     else {
-        half v = __float2half(grad_output[blockIdx.x]);
+        nv_bfloat16 v = __float2bfloat16(grad_output[blockIdx.x]);
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             grad_input[base_idx + i] = i==t ? __hsub(__hmul(softmax[base_idx + i], v), v) : __hmul(softmax[base_idx + i], v);
         }
@@ -70,7 +71,7 @@ __global__ void cross_entropy_backward(
 // blocks <m>,      threads<1024>
 __global__ void cross_entropy_forward_inplace(
     int64_t n,
-    half *x,                // (m, n)
+    nv_bfloat16 *x,                // (m, n)
     const int32_t *target,  // (m)
     float *output,          // (m)
     int32_t ignore_index
@@ -79,19 +80,19 @@ __global__ void cross_entropy_forward_inplace(
 
     float local_max = -INFINITY;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_max = fmaxf(__half2float(x[base_idx + i]), local_max);
+        local_max = fmaxf(__bfloat162float(x[base_idx + i]), local_max);
     }
     local_max = fmaxf(block_allreduce_max(local_max), -1e6);
     
     float local_sum = 0;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_sum += expf(__half2float(x[base_idx + i]) - local_max);
+        local_sum += expf(__bfloat162float(x[base_idx + i]) - local_max);
     }
     local_sum = block_allreduce_sum(local_sum) + 1e-10; // avoid nan
 
     if (threadIdx.x == 0) {
         if (target[blockIdx.x] != ignore_index) {
-            output[blockIdx.x] = -__half2float(x[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
+            output[blockIdx.x] = -__bfloat162float(x[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
         } else {
             output[blockIdx.x] = 0;
         }
@@ -100,7 +101,7 @@ __global__ void cross_entropy_forward_inplace(
     __syncthreads();
     
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        x[base_idx + i] = __float2half( expf(__half2float(x[base_idx + i]) - local_max) / local_sum );
+        x[base_idx + i] = __float2bfloat16( expf(__bfloat162float(x[base_idx + i]) - local_max) / local_sum );
     }
 }
 
@@ -109,20 +110,20 @@ __global__ void cross_entropy_backward_inplace(
     int64_t n,
     const float *grad_output,   // (m)
     const int32_t *target,      // (m)
-    half *x,                    // (m, n)
+    nv_bfloat16 *x,                    // (m, n)
     int32_t ignore_index
 ) {
     int64_t base_idx = blockIdx.x * n;
 
     int32_t t = target[blockIdx.x];
     if (t == ignore_index) {
-        half v = __float2half(0.);
+        nv_bfloat16 v = __float2bfloat16(0.);
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             x[base_idx + i] = v;
         }
     }
     else {
-        half v = __float2half(grad_output[blockIdx.x]);
+        nv_bfloat16 v = __float2bfloat16(grad_output[blockIdx.x]);
         __syncthreads();
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             x[base_idx + i] = i==t ? __hsub(__hmul(x[base_idx + i], v), v) : __hmul(x[base_idx + i], v);
@@ -140,9 +141,9 @@ void cross_entropy_forward_launcher(
     torch::Tensor &output,
     int32_t ignore_index
 ) {
-    auto input_ptr = reinterpret_cast<half*>(input.data_ptr<at::Half>());
+    auto input_ptr = reinterpret_cast<nv_bfloat16*>(input.data_ptr<at::BFloat16>());
     auto target_ptr = target.data_ptr<int32_t>();
-    auto softmax_ptr = reinterpret_cast<half*>(softmax.data_ptr<at::Half>());
+    auto softmax_ptr = reinterpret_cast<nv_bfloat16*>(softmax.data_ptr<at::BFloat16>());
     auto output_ptr = output.data_ptr<float>();
     int32_t threads = 1024;
     auto stream = at::cuda::getCurrentCUDAStream();
@@ -159,8 +160,8 @@ void cross_entropy_backward_launcher(
 ) {
     auto output_ptr = grad_output.data_ptr<float>();
     auto target_ptr = target.data_ptr<int32_t>();
-    auto softmax_ptr = reinterpret_cast<half*>(softmax.data_ptr<at::Half>());
-    auto input_ptr = reinterpret_cast<half*>(grad_input.data_ptr<at::Half>());
+    auto softmax_ptr = reinterpret_cast<nv_bfloat16*>(softmax.data_ptr<at::BFloat16>());
+    auto input_ptr = reinterpret_cast<nv_bfloat16*>(grad_input.data_ptr<at::BFloat16>());
     int32_t threads = 1024;
     auto stream = at::cuda::getCurrentCUDAStream();
     cross_entropy_backward<<<m, threads, 0, stream.stream()>>>(n, output_ptr, target_ptr, softmax_ptr, input_ptr, ignore_index);
@@ -173,7 +174,7 @@ void cross_entropy_forward_inplace_launcher(
     torch::Tensor &output,
     int32_t ignore_index
 ) {
-    auto x_ptr = reinterpret_cast<half*>(x.data_ptr<at::Half>());
+    auto x_ptr = reinterpret_cast<nv_bfloat16*>(x.data_ptr<at::BFloat16>());
     auto target_ptr = target.data_ptr<int32_t>();
     auto output_ptr = output.data_ptr<float>();
     int32_t threads = 1024;
@@ -190,7 +191,7 @@ void cross_entropy_backward_inplace_launcher(
 ) {
     auto output_ptr = grad_output.data_ptr<float>();
     auto target_ptr = target.data_ptr<int32_t>();
-    auto x_ptr = reinterpret_cast<half*>(x.data_ptr<at::Half>());
+    auto x_ptr = reinterpret_cast<nv_bfloat16*>(x.data_ptr<at::BFloat16>());
     int32_t threads = 1024;
     auto stream = at::cuda::getCurrentCUDAStream();
     cross_entropy_backward_inplace<<<m, threads, 0, stream.stream()>>>(n, output_ptr, target_ptr, x_ptr, ignore_index);
